Add ability to add stocks & major morph (#134)
* Add UI.print() method

No functional changes yet.

* Shift Parse.parse() exception printing to UI

Cascaded all Parse.parse() exceptions upwards to be printed by UI.printError(Exception e) within Duke.run()

* Add UI output indentation

The idea is that UI.print(String printString) will automatically handle printing lines outside every output by the program, and will also indent all output to clearly differentiate between user input and program output.

* Extract TaskList.searchForTask to FindCommand.execute()

We want to make TaskList just a dumb container, any smarter processing should be done within the Command objects' execute() methods.

* Extract printing of FindCommand to UI.print()

* Extract TaskList.printList() to Command.execute()

Ideally I believe we should have an explicit ListCommand class that executes this, but for now this should make dealing with UI output easier.

* Add DoneCommand class

Also extract TaskList.markItemAsDone() method

* Add DeleteCommand class

Also extract TaskList.deleteTask() method

* Add explicit ListCommand class

* Polish UI.print() extraction refactor

* Fix .gitignore for saved_tasks.txt

* Fix file permissions

* Fix travis checkstyle 1

* Fix travis checkstyle 2

ughhhh

* Disable JUnit test for NumCommand

* Disable JUnit test for FindCommand & printList

Disabled temporarily after refactor

* Rename duke occurrences to eggventory

* Rename TaskList to StockType

* Finish initial morph from Tasks to Stocks

stonks*

* index on stock-morph: ea60fa9 Finish initial morph from Tasks to Stocks

* Fix FindCommand

It's a very simplistic implementation right now.

* Fix .* import statement checkstyle

* Fix typo
