+++ build.gradle
@@ -11,12 +11,12 @@ checkstyle {
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "eggventory"
     archiveVersion = "0.1.3"
     archiveClassifier = null
     archiveAppendix = null
 }
-group 'seedu.duke'
+group 'seedu.eggventory'
 version '0.1.0'
 
 
@@ -38,7 +38,7 @@ repositories {
 
 application {
     // Change this to your main class.
-    mainClassName = "duke.Duke"
+    mainClassName = "eggventory.Duke"
 }
 
 run {
