+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'duke'
+rootProject.name = 'eggventory'
