--- src/main/java/duke/EventDateTimeComparator.java
@@ -1,10 +1,0 @@
-package duke;
-
-import duke.items.Event;
-import java.util.Comparator;
-
-public class EventDateTimeComparator implements Comparator<Event> {
-    public int compare(Event first, Event second) {
-        return first.getEventStartTimeObj().getAt().compareTo(second.getEventStartTimeObj().getAt());
-    }
-}
