--- src/main/java/duke/commands/DoneCommand.java
@@ -1,27 +1,0 @@
-package duke.commands;
-
-import duke.TaskList;
-import duke.Ui;
-import duke.Storage;
-import duke.enums.CommandType;
-
-public class DoneCommand extends Command {
-    private int itemIndex;
-
-    public DoneCommand(CommandType type, int index) {
-        super(type);
-        this.itemIndex = index - 1;
-    }
-
-    @Override
-    public void execute(TaskList list, Ui ui, Storage storage) throws IndexOutOfBoundsException {
-        try {
-            list.getTask(itemIndex).markAsDone();
-            ui.print("Nice! I've marked this task as done:\n"
-                    + list.getTask(itemIndex).toString());
-        } catch (IndexOutOfBoundsException e) {
-            throw new IndexOutOfBoundsException("That task doesn't exist! Please check"
-                    + " the available tasks again: ");
-        }
-    }
-}
