--- src/main/java/duke/enums/CommandType.java
@@ -1,5 +1,0 @@
-package duke.enums;
-
-public enum CommandType {
-    TODO, DEADLINE, EVENT, BYE, LIST, REMINDER, DONE, DELETE, FIND, BAD
-}
