--- src/main/java/duke/enums/TaskType.java
@@ -1,5 +1,0 @@
-package duke.enums;
-
-public enum TaskType {
-    TODO, DEADLINE, EVENT
-}
