--- src/main/java/duke/exceptions/BadInputException.java
@@ -1,8 +1,0 @@
-package duke.exceptions;
-
-public class BadInputException extends Exception {
-
-    public BadInputException(String s) {
-        super(s);
-    }
-}
