--- src/main/java/duke/exceptions/DukeExceptions.java
@@ -1,8 +1,0 @@
-package duke.exceptions;
-/**
- * Unsure of how to use DukeExceptions properly.
- */
-
-public class DukeExceptions extends Exception {
-
-}
