--- src/main/java/duke/exceptions/InsufficientInfoException.java
@@ -1,9 +1,0 @@
-package duke.exceptions;
-
-public class InsufficientInfoException extends Exception {
-
-    public InsufficientInfoException(String s) {
-        super(s);
-    }
-
-}
