+++ src/main/java/eggventory/enums/CommandType.java
@@ -1,0 +1,5 @@
+package eggventory.enums;
+
+public enum CommandType {
+    ADD, TODO, DEADLINE, EVENT, BYE, LIST, REMINDER, DONE, DELETE, FIND, BAD
+}
