+++ src/main/java/eggventory/enums/TaskType.java
@@ -1,0 +1,5 @@
+package eggventory.enums;
+
+public enum TaskType {
+    TODO, DEADLINE, EVENT
+}
