+++ src/main/java/eggventory/exceptions/BadInputException.java
@@ -1,0 +1,8 @@
+package eggventory.exceptions;
+
+public class BadInputException extends Exception {
+
+    public BadInputException(String s) {
+        super(s);
+    }
+}
