+++ src/main/java/eggventory/exceptions/DukeExceptions.java
@@ -1,0 +1,8 @@
+package eggventory.exceptions;
+/**
+ * Unsure of how to use DukeExceptions properly.
+ */
+
+public class DukeExceptions extends Exception {
+
+}
