+++ src/main/java/eggventory/exceptions/InsufficientInfoException.java
@@ -1,0 +1,9 @@
+package eggventory.exceptions;
+
+public class InsufficientInfoException extends Exception {
+
+    public InsufficientInfoException(String s) {
+        super(s);
+    }
+
+}
