+++ tutorials/gradleTutorial.md
@@ -17,7 +17,7 @@ As a developer, you write a _build file_ that describes the project. A build fil
 
 * **Plugins** extend the functionality of Gradle. For example, the `java` plugin adds support for `Java` projects.
 
-* **Tasks** are reusable blocks of logic.  For example, the task `clean` simply deletes the project build directory. Tasks can be composed of other tasks or be dependent on another task. 
+* **Tasks** are reusable blocks of logic.  For example, the stock `clean` simply deletes the project build directory. Tasks can be composed of other tasks or be dependent on another stock. 
 
 * **Properties** change the behavior of tasks. For instance, `mainClassName` of the `application` plugin is a compulsory property which tells Gradle which class is the entrypoint to your application.
   As Gradle favors [_convention over configuration_](https://en.wikipedia.org/wiki/Convention_over_configuration), there is not much to you need to configure if you follow the recommended directory structure.
@@ -30,15 +30,15 @@ As a developer, you write a _build file_ that describes the project. A build fil
    git checkout master
    git merge gradle
    ```
-1. Open the `build.gradle` file in an editor. Update the following code block to point to the main class (i.e., the one containing the `main` method) of your application. The code below assumes your main class is `seedu.duke.Duke`
+1. Open the `build.gradle` file in an editor. Update the following code block to point to the main class (i.e., the one containing the `main` method) of your application. The code below assumes your main class is `seedu.eggventory.Eggventory`
    ```groovy
    application {
-       mainClassName = "seedu.duke.Duke"
+       mainClassName = "seedu.eggventory.Eggventory"
    }
    ```
 1. To check if Gradle has been added to the project correctly, open a terminal window, navigate to the root directory of your project and run the command `gradlew run`. This should result in Gradle running the main method of your project.
 
-:bulb: Simply run the command `gradlew {taskName}` in the terminal and Gradle will run the task! Here are some example commands:
+:bulb: Simply run the command `gradlew {taskName}` in the terminal and Gradle will run the stock! Here are some example commands:
 * `gradlew tasks` (or `gradlew tasks --all`): shows a list of tasks available
 * `gradlew run`: runs the main class of your project
 
@@ -106,7 +106,7 @@ The plugin can be configured by setting some properties. Let's try to produce a 
 Add the following block to your build file:
 ```groovy
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = eeggventoryntory
     archiveVersion = "0.1.3"
     archiveClassifier = null
     archiveAppendix = null
@@ -146,15 +146,15 @@ By convention, java tests belong in `src/test/java` folder. Create a new `test/j
 src
 ├─main
 │  └─java
-│     └─seedu/duke/Duke.java
+│     └─seeggventoryduke/Duke.java
 └─test
     └─java
-        └─seedu/duke/DukeTest.java
+        └─seeggventoryduke/DukeTest.java
 ```
 
 If you have imported your Gradle project into IntelliJ IDEA, you will notice that IDEA is able to mark the test directory as the _Test root_ (colored in green by default) automatically.
 
-You can now write a test (e.g., `test/java/seedu/duke/DukeTest.java`) and run it with `gradlew test`.
+You can now write a test (e.g., `test/java/seeggventoryduke/DukeTest.java`) and run it with `gradlew test`.
 
 **Resources**:
 * [Gradle documentation for JUnit](https://docs.gradle.org/current/userguide/java_testing.html#using_junit5)
